+++ packages/types/src/interfaces/contractsAbi/types.ts
@@ -74,6 +74,9 @@ export interface ContractContractSpecV3 extends Struct {
   readonly docs: Vec<Text>;
 }
 
+/** @name ContractContractSpecV4 */
+export interface ContractContractSpecV4 extends ContractContractSpecV3 {}
+
 /** @name ContractCryptoHasher */
 export interface ContractCryptoHasher extends Enum {
   readonly isBlake2x256: boolean;
@@ -240,11 +243,13 @@ export interface ContractMetadata extends Enum {
   readonly asV2: ContractMetadataV2;
   readonly isV3: boolean;
   readonly asV3: ContractMetadataV3;
-  readonly type: 'V0' | 'V1' | 'V2' | 'V3';
+  readonly isV4: boolean;
+  readonly asV4: ContractMetadataV4;
+  readonly type: 'V0' | 'V1' | 'V2' | 'V3' | 'V4';
 }
 
 /** @name ContractMetadataLatest */
-export interface ContractMetadataLatest extends ContractMetadataV3 {}
+export interface ContractMetadataLatest extends ContractMetadataV4 {}
 
 /** @name ContractMetadataV0 */
 export interface ContractMetadataV0 extends Struct {
@@ -267,6 +272,12 @@ export interface ContractMetadataV2 extends Struct {
 
 /** @name ContractMetadataV3 */
 export interface ContractMetadataV3 extends Struct {
+  readonly types: Vec<PortableType>;
+  readonly spec: ContractContractSpecV3;
+}
+
+/** @name ContractMetadataV4 */
+export interface ContractMetadataV4 extends Struct {
   readonly types: Vec<PortableType>;
   readonly spec: ContractContractSpecV3;
 }