+++ packages/types/src/interfaces/runtime/definitions.ts
@@ -53,7 +53,8 @@ export default {
   types: objectSpread({}, numberTypes, {
     AccountId: 'AccountId32',
     AccountId20: 'GenericEthereumAccountId',
-    AccountId32: 'GenericAccountId',
+    AccountId32: 'GenericAccountId32',
+    AccountId33: 'GenericAccountId33',
     AccountIdOf: 'AccountId',
     AccountIndex: 'GenericAccountIndex',
     Address: 'MultiAddress',