+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 Changes:
 
+- Support for ink! v4 metadata
 - Adjust `pendingRewards` for pools runtime calls
 - Update to latest Substrate, Polkadot & Kusama metadata