+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 Changes:
 
 - Adjust `pendingRewards` for pools runtime calls
+- Support for `AccountId33` (raw/compressed 33-byte ecdsa)
 - Update to latest Substrate, Polkadot & Kusama metadata