+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 Changes:
 
 - Support for ink! v4 metadata
-- Adjust `pendingRewards` for pools runtime calls
+- Adjust pool runtime call `pendingRewards` return type
 - Support for `AccountId33` (raw/compressed 33-byte ecdsa)
 - Update to latest Substrate, Polkadot & Kusama metadata