+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG
 
-## master
+## 9.3.1 Sep 3, 2022
 
 Changes: